+++ internal/app/wwctl/container/exec/root.go
@@ -26,6 +26,7 @@ var (
 		FParseErrWhitelist: cobra.FParseErrWhitelist{UnknownFlags: true},
 	}
 	SyncUser bool
+	Build    bool
 	binds    []string
 	nodeName string
 )
@@ -37,6 +38,7 @@ Bind a local path which must exist into the container. If destination is not
 set, uses the same path as source. "ro" binds read-only. "copy" temporarily
 copies the file into the container.`)
 	baseCmd.PersistentFlags().BoolVar(&SyncUser, "syncuser", false, "Synchronize UIDs/GIDs from host to container")
+	baseCmd.PersistentFlags().BoolVar(&Build, "build", true, "(Re)build the container image automatically")
 	baseCmd.PersistentFlags().StringVarP(&nodeName, "node", "n", "", "Create a read only view of the container for the given node")
 }