+++ CHANGELOG.md
@@ -33,6 +33,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Support defining a symlink with an overlay template. #1303
 - New "localtime" overlay to define the system time zone. #1303
 - Add support for nested profiles. #1572, #1598
+- Adds `wwctl container <exec|shell> --build=false` to prevent automatically (re)building the container. #1490, #1489
 
 ### Changed